Change circular buffer to float and fix multiple tasks using FPU

+++ circularBuffer.h
@@ -4,7 +4,7 @@
 #include "mutex.h"
 
 typedef struct {
-	int data[10];
+	float data[10];
 	int insert;
 	int retrieve;
 	int isFull;
@@ -12,7 +12,7 @@ typedef struct {
 } queue_t;
 
 void queue_init(queue_t* queue);
-int queue_put(queue_t* queue, int item);
-int queue_get(queue_t* queue, int* item);
+int queue_put(queue_t* queue, float item);
+int queue_get(queue_t* queue, float* item);
 
 #endif /* __circularBuffer_h__ */
